removed NewNegotiation and converted int to float for prices

+++ ClientHandler/src/main/java/ir/sharif/gamein2021/ClientHandler/domain/RFQ/NewProviderNegotiationRequest.java
@@ -13,5 +13,5 @@ public class NewProviderNegotiationRequest extends RequestObject implements Seri
 {
     private Integer providerId;
     private Integer amount;
-    private Integer costPerUnitDemander;
+    private Float costPerUnitDemander;
 }
