+++ Core/src/main/java/ir/sharif/gamein2021/core/domain/dto/NegotiationDto.java
@@ -18,8 +18,8 @@ public class NegotiationDto implements BaseDto<Integer> {
     private Integer supplierId;
     private Integer productId;
     private Integer amount;
-    private Integer costPerUnitDemander;
-    private Integer costPerUnitSupplier;
+    private Float costPerUnitDemander;
+    private Float costPerUnitSupplier;
     private NegotiationState state;
 
 }
